requirements.txt
transformers
pyYAML
python-dotenv
av
aiortc
langchain
streamlit-webrtc
streamlit 
python-docx
python-pptx
PyPDF2
langchain-huggingface
langchain-google-genai
langchain-pinecone
langchain-core
pinecone-client
moviepy